Introduce shared ansible inventory processing
