+++ app/models/manageiq/providers/ansible_tower/inventory/collector/automation_manager.rb
@@ -1,25 +1,3 @@
 class ManageIQ::Providers::AnsibleTower::Inventory::Collector::AutomationManager < ManagerRefresh::Inventory::Collector
-  def connection
-    @connection ||= manager.connect
-  end
-
-  def inventories
-    connection.api.inventories.all
-  end
-
-  def hosts
-    connection.api.hosts.all
-  end
-
-  def job_templates
-    connection.api.job_templates.all
-  end
-
-  def projects
-    connection.api.projects.all
-  end
-
-  def credentials
-    connection.api.credentials.all
-  end
+  include ManageIQ::Providers::AnsibleTower::Shared::Inventory::Collector::AutomationManager
 end
