+++ app/models/manageiq/providers/embedded_ansible/inventory/collector/automation_manager.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAnsible::Inventory::Collector::AutomationManager < ManagerRefresh::Inventory::Collector
+  include ManageIQ::Providers::AnsibleTower::Shared::Inventory::Collector::AutomationManager
+end
