+++ app/models/manageiq/providers/embedded_ansible/inventory/parser/automation_manager.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAnsible::Inventory::Parser::AutomationManager < ManagerRefresh::Inventory::Parser
+  include ManageIQ::Providers::AnsibleTower::Shared::Inventory::Parser::AutomationManager
+end
