+++ app/models/manageiq/providers/embedded_ansible/inventory/persister/automation_manager.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAnsible::Inventory::Persister::AutomationManager < ManagerRefresh::Inventory::Persister
+  include ManageIQ::Providers::AnsibleTower::Shared::Inventory::Persister::AutomationManager
+end
