+++ app/models/manageiq/providers/embedded_ansible/inventory_collection_default/automation_manager.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAnsible::InventoryCollectionDefault::AutomationManager < ManagerRefresh::InventoryCollectionDefault
+  include ManageIQ::Providers::AnsibleTower::Shared::InventoryCollectionDefault::AutomationManager
+end
